stuck, trying to make content banners span 100 per cent of page and be of equal h and w
I want content banners 1 & 2 to be the same width and height.  I want them to be side by side in desktop and table mode, and on top of each other in cell mode.
Frustrated because I cannot get them to span 100% of screen.  Also cannot get them to be of equal width.
Note: in previous iteration I cleaned up/condensed the code around content banners, removed a <div> and a style.